Add files via upload
This is the data file which is required in running the code mentioned as "project.R". It has abstract, title, author and other details of the selected academic corpora on British mammals. The file has necessary information to run a topic model on the literature published regarding  terrestrial mammals of Britain